+++ src/test_cases/test_case_4_material_b.rs
@@ -20,7 +20,7 @@ use crate::ray_tracing::primitives::sphere::Sphere;
 use crate::ray_tracing::renderer::Renderer;
 use crate::ray_tracing::world::World;
 
-pub fn test(samples: i32) {
+pub fn test(samples: usize) {
     let file_name = get_file_name(file!());
     println!("TESTING: {}", &file_name);
     let ppm_name = format!("{}.ppm", file_name);
